src/app/requirements.txt
web.py==0.40
mysql-connector-python==8.0.19
python-dotenv
bcrypt